revise employeeController, organize directories
