--- employee-details-frontend/src/AddEmployeeForm.css
@@ -1,35 +1,0 @@
-.popup {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(0, 0, 0, 0.2);
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-#add-employee-form {
-  position: relative;
-  padding: 32px;
-  width: 60vw;
-  background-color: white;
-
-  display: flex;
-  flex-direction: column;
-}
-
-.add-employee-input {
-  margin-bottom: 15px;
-  width: 60vw
-}
-
-#add-employee-buttons {
-  padding-top: 10px;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-}
