+++ employee-details-frontend/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import EmployeeList from './EmployeeList'; // Make sure the path is correct
-import UploadPage from './UploadPage'; // Make sure the path is correct
+import UploadPage from './pages/UploadPage'; // Make sure the path is correct
 
 function App() {
     return (
